Add page_generator class with functional website
