+++ site/index.html
@@ -1,0 +1,78 @@
+
+    <html lang="en-US">
+      <head>
+        <meta charset = "UTF-8">
+        <title>Futbol Statistics</title>
+        <link rel="stylesheet" href="stylesheet.css" />
+      </head>
+
+      <body>
+        <h1>Futbol Statistics</h1>
+        <hr>
+
+        <h3>Game Statistics</h3>
+
+        <!--
+        <ul>
+          
+            <li>{"team_id"=>"1", "franchise_id"=>"23", "team_name"=>"Atlanta United", "abbreviation"=>"ATL", "link"=>"/api/v1/teams/1"}</li>
+          
+            <li>{"team_id"=>"4", "franchise_id"=>"16", "team_name"=>"Chicago Fire", "abbreviation"=>"CHI", "link"=>"/api/v1/teams/4"}</li>
+          
+            <li>{"team_id"=>"6", "franchise_id"=>"6", "team_name"=>"FC Dallas", "abbreviation"=>"DAL", "link"=>"/api/v1/teams/6"}</li>
+          
+            <li>{"team_id"=>"3", "franchise_id"=>"10", "team_name"=>"Houston Dynamo", "abbreviation"=>"HOU", "link"=>"/api/v1/teams/3"}</li>
+          
+            <li>{"team_id"=>"28", "franchise_id"=>"29", "team_name"=>"Los Angeles FC", "abbreviation"=>"LFC", "link"=>"/api/v1/teams/28"}</li>
+          
+            <li>{"team_id"=>"16", "franchise_id"=>"11", "team_name"=>"New England Revolution", "abbreviation"=>"NE", "link"=>"/api/v1/teams/16"}</li>
+          
+            <li>{"team_id"=>"30", "franchise_id"=>"37", "team_name"=>"Orlando City SC", "abbreviation"=>"ORL", "link"=>"/api/v1/teams/30"}</li>
+          
+            <li>{"team_id"=>"15", "franchise_id"=>"24", "team_name"=>"Portland Timbers", "abbreviation"=>"POR", "link"=>"/api/v1/teams/15"}</li>
+          
+            <li>{"team_id"=>"19", "franchise_id"=>"18", "team_name"=>"Philadelphia Union", "abbreviation"=>"PHI", "link"=>"/api/v1/teams/19"}</li>
+          
+            <li>{"team_id"=>"24", "franchise_id"=>"32", "team_name"=>"Real Salt Lake", "abbreviation"=>"RSL", "link"=>"/api/v1/teams/24"}</li>
+          
+            <li>{"team_id"=>"27", "franchise_id"=>"28", "team_name"=>"San Jose Earthquakes", "abbreviation"=>"SJ", "link"=>"/api/v1/teams/27"}</li>
+          
+            <li>{"team_id"=>"2", "franchise_id"=>"22", "team_name"=>"Seattle Sounders FC", "abbreviation"=>"SEA", "link"=>"/api/v1/teams/2"}</li>
+          
+            <li>{"team_id"=>"20", "franchise_id"=>"21", "team_name"=>"Toronto FC", "abbreviation"=>"TOR", "link"=>"/api/v1/teams/20"}</li>
+          
+            <li>{"team_id"=>"21", "franchise_id"=>"27", "team_name"=>"Vancouver Whitecaps FC", "abbreviation"=>"VAN", "link"=>"/api/v1/teams/21"}</li>
+          
+            <li>{"team_id"=>"13", "franchise_id"=>"33", "team_name"=>"Houston Dash", "abbreviation"=>"HOD", "link"=>"/api/v1/teams/13"}</li>
+          
+            <li>{"team_id"=>"10", "franchise_id"=>"5", "team_name"=>"North Carolina Courage", "abbreviation"=>"NCC", "link"=>"/api/v1/teams/10"}</li>
+          
+            <li>{"team_id"=>"29", "franchise_id"=>"36", "team_name"=>"Orlando Pride", "abbreviation"=>"FLP", "link"=>"/api/v1/teams/29"}</li>
+          
+            <li>{"team_id"=>"52", "franchise_id"=>"35", "team_name"=>"Portland Thorns FC", "abbreviation"=>"ORT", "link"=>"/api/v1/teams/52"}</li>
+          
+            <li>{"team_id"=>"12", "franchise_id"=>"26", "team_name"=>"Sky Blue FC", "abbreviation"=>"SBL", "link"=>"/api/v1/teams/12"}</li>
+          
+            <li>{"team_id"=>"22", "franchise_id"=>"25", "team_name"=>"Washington Spirit FC", "abbreviation"=>"WAS", "link"=>"/api/v1/teams/22"}</li>
+          
+        </ul> -->
+
+        <h4>Average Goals by Season:</h4>
+        <table>
+          <tr>
+            <th>Season</th>
+            <th>Average Goals</th>
+          </tr>
+          <tr>
+            <td>Jill</td>
+            <td>Smith</td>
+          </tr>
+          <tr>
+            <td>Eve</td>
+            <td>Jackson</td>
+          </tr>
+        </table>
+
+      </body>
+    </html>
+
