+++ site/stylesheet.css
@@ -1,0 +1,3 @@
+table, th, td {
+  border: 1px solid black;
+}
